+++ src/interfaces/mod.rs
@@ -8,4 +8,4 @@ pub mod pyo3;
 
 #[cfg(feature = "wasm_bindgen")]
 pub mod wasm_bindgen;
-pub use crate::core::policy;
+pub use abe_policy;